Create RockPaperScissorGame.py
This is a basic game made with python. Here you have to choose one of the given characters which are rock, paper, and scissor. I used random module for copmuter. I put the whole game into a while loop so that if you want to play again you don't have to run the program again. And you can quit or restart the game simply by typing quit or restart.